+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/Route.kt
@@ -1,6 +1,7 @@
 package net.nymtech.nymvpn.ui
 
 import kotlinx.serialization.Serializable
+import nym_vpn_lib_types.GatewayType
 
 sealed class Route {
 	@Serializable
@@ -17,6 +18,9 @@ sealed class Route {
 
 	@Serializable
 	data class Settings(val showVpnSettings: Boolean) : Route()
+
+	@Serializable
+	data object Censorship : Route()
 
 	@Serializable
 	data object Appearance : Route()
@@ -62,4 +66,11 @@ sealed class Route {
 
 	@Serializable
 	data object SelectPlan : Route()
+
+	@Serializable
+	data class ServerDetails(
+		val id: String,
+		val type: GatewayType,
+		val location: String,
+	) : Route()
 }