+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/common/navigation/NavBar.kt
@@ -121,6 +121,15 @@ fun NavBar(navController: NavController, modifier: Modifier = Modifier) {
 					}
 				},
 			)
+			currentRoute.startsWith(Route.Censorship::class.qualifiedName!!) -> NavBarState(
+				title = { NavTitle(stringResource(R.string.settings_censorship_title)) },
+				show = true,
+				leading = {
+					NavIcon(Icons.AutoMirrored.Filled.ArrowBack, stringResource(R.string.back)) {
+						navController.safePopBackStack()
+					}
+				},
+			)
 			currentRoute.startsWith(Route.Appearance::class.qualifiedName!!) -> NavBarState(
 				title = { NavTitle(stringResource(R.string.appearance)) },
 				show = true,
@@ -131,7 +140,7 @@ fun NavBar(navController: NavController, modifier: Modifier = Modifier) {
 				},
 			)
 			currentRoute.startsWith(Route.Privacy::class.qualifiedName!!) -> NavBarState(
-				title = { NavTitle(stringResource(R.string.data_privacy)) },
+				title = { NavTitle(stringResource(R.string.privacy_title)) },
 				show = true,
 				leading = {
 					NavIcon(Icons.AutoMirrored.Filled.ArrowBack, stringResource(R.string.back)) {
@@ -178,6 +187,15 @@ fun NavBar(navController: NavController, modifier: Modifier = Modifier) {
 			currentRoute.startsWith(Route.LoginScanner::class.qualifiedName!!) -> NavBarState(show = false)
 			currentRoute.startsWith(Route.Permission::class.qualifiedName!!) -> NavBarState(
 				title = { NavTitle(stringResource(R.string.permission_required)) },
+				show = true,
+				leading = {
+					NavIcon(Icons.AutoMirrored.Filled.ArrowBack, stringResource(R.string.back)) {
+						navController.safePopBackStack()
+					}
+				},
+			)
+			currentRoute.startsWith(Route.ServerDetails::class.qualifiedName!!) -> NavBarState(
+				title = { NavTitle(stringResource(R.string.details_title)) },
 				show = true,
 				leading = {
 					NavIcon(Icons.AutoMirrored.Filled.ArrowBack, stringResource(R.string.back)) {