+++ nym-vpn-android/CHANGELOG.md
@@ -6,10 +6,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.1.0]
+
+### Added
 - Themed icons support (https://github.com/nymtech/nym-vpn-client/pull/3429)
-- UI state updates after logout (https://github.com/nymtech/nym-vpn-client/pull/3448)
 - Random option added for EntryPoint (https://github.com/nymtech/nym-vpn-client/pull/3448)
+- Server details screen (https://github.com/nymtech/nym-vpn-client/pull/3472)
+- UI for Domain fronting (https://github.com/nymtech/nym-vpn-client/pull/3490)
+- UI for QUIC (https://github.com/nymtech/nym-vpn-client/pull/3490)
+- Feature flag support (https://github.com/nymtech/nym-vpn-client/pull/3490)
+
+### Changed
 - Detailed info status for Connecting state (https://github.com/nymtech/nym-vpn-client/pull/3448)
+- Server name now displayed below Country on Main Screen (https://github.com/nymtech/nym-vpn-client/pull/3465)
+
+### Fixed
+- UI state updates after logout (https://github.com/nymtech/nym-vpn-client/pull/3448)
+- UI fixes for (https://github.com/nymtech/nym-vpn-client/pull/3490)
 
 ## [2.0.0] - 2025-09-10